+++ packages/gamut/src/Form/styles/Checkbox-styles.ts
@@ -67,7 +67,7 @@ export const checkboxElementStates = system.states({
   multiline: {
     mt: 4,
   },
-  checked: {
+  active: {
     color: 'primary',
   },
   disabled: {
@@ -75,6 +75,12 @@ export const checkboxElementStates = system.states({
     [InputSelectors.HOVER]: {
       outline: 'none',
     },
+  },
+  hasBg: {
+    bg: 'currentColor',
+  },
+  hideBorder: {
+    borderColor: 'transparent',
   },
 });