+++ packages/gamut-styles/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [17.8.0](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-styles@17.7.0...@codecademy/gamut-styles@17.8.0) (2025-08-20)
+
+### Features
+
+- add nonce property to GamutProvider [FCORE-5007] ([190e36f](https://github.com/Codecademy/gamut/commit/190e36f98de6543515bb32f68054d5a26526d863))
+
+## [17.7.0](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-styles@17.6.5...@codecademy/gamut-styles@17.7.0) (2025-07-22)
+
+### Features
+
+- **ContainerQueries:** add container queries to responsive props api ([f29d307](https://github.com/Codecademy/gamut/commit/f29d3071599f9c686b47687787ed44d1e9d826e1))
+
+### [17.6.5](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-styles@17.6.4...@codecademy/gamut-styles@17.6.5) (2025-07-17)
+
+**Note:** Version bump only for package @codecademy/gamut-styles
+
 ### [17.6.4](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-styles@17.6.3...@codecademy/gamut-styles@17.6.4) (2025-06-17)
 
 **Note:** Version bump only for package @codecademy/gamut-styles